moved search to search.py, added dask
